# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qlib Trading Platform API",
    "description": "Quantitative trading with Qlib research engine",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/broker/orders": {
      "get": {
        "tags": [
          "broker"
        ],
        "summary": "List Orders",
        "description": "Get recent orders for current user.",
        "operationId": "list_orders_api_broker_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderResponse"
                  },
                  "type": "array",
                  "title": "Response List Orders Api Broker Orders Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "broker"
        ],
        "summary": "Place Order",
        "description": "Place an order and persist it.",
        "operationId": "place_order_api_broker_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/broker/orders/{order_id}/execute": {
      "post": {
        "tags": [
          "broker"
        ],
        "summary": "Execute Pending Order",
        "description": "Execute pending limit/stop order at current price.",
        "operationId": "execute_pending_order_api_broker_orders__order_id__execute_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          },
          {
            "name": "current_price",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Current Price"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/broker/orders/{order_id}": {
      "get": {
        "tags": [
          "broker"
        ],
        "summary": "Get Order",
        "description": "Get order status by id.",
        "operationId": "get_order_api_broker_orders__order_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/broker/portfolio": {
      "get": {
        "tags": [
          "broker"
        ],
        "summary": "Get Portfolio",
        "description": "Get current portfolio state.",
        "operationId": "get_portfolio_api_broker_portfolio_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "current_prices",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON string map of symbol->price",
              "title": "Current Prices"
            },
            "description": "JSON string map of symbol->price"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/broker/positions": {
      "get": {
        "tags": [
          "broker"
        ],
        "summary": "Get Positions",
        "description": "Get all open positions.",
        "operationId": "get_positions_api_broker_positions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "current_prices",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON string map of symbol->price",
              "title": "Current Prices"
            },
            "description": "JSON string map of symbol->price"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionResponse"
                  },
                  "title": "Response Get Positions Api Broker Positions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/broker/positions/{ticker}": {
      "get": {
        "tags": [
          "broker"
        ],
        "summary": "Get Position",
        "description": "Get single position by ticker.",
        "operationId": "get_position_api_broker_positions__ticker__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "current_prices",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON string map of symbol->price",
              "title": "Current Prices"
            },
            "description": "JSON string map of symbol->price"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/broker/trades/closed": {
      "get": {
        "tags": [
          "broker"
        ],
        "summary": "Get Closed Trades",
        "description": "Get all persisted trades.",
        "operationId": "get_closed_trades_api_broker_trades_closed_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClosedTradeResponse"
                  },
                  "type": "array",
                  "title": "Response Get Closed Trades Api Broker Trades Closed Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/broker/statistics": {
      "get": {
        "tags": [
          "broker"
        ],
        "summary": "Get Statistics",
        "description": "Get persisted portfolio statistics.",
        "operationId": "get_statistics_api_broker_statistics_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "current_prices",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON string map of symbol->price",
              "title": "Current Prices"
            },
            "description": "JSON string map of symbol->price"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/broker/reset": {
      "post": {
        "tags": [
          "broker"
        ],
        "summary": "Reset Portfolio",
        "description": "Reset default portfolio and all associated trades/orders/positions.",
        "operationId": "reset_portfolio_api_broker_reset_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClosedTradeResponse": {
        "properties": {
          "trade_id": {
            "type": "string",
            "title": "Trade Id"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "side": {
            "type": "string",
            "title": "Side"
          },
          "entry_date": {
            "type": "string",
            "title": "Entry Date"
          },
          "entry_price": {
            "type": "number",
            "title": "Entry Price"
          },
          "entry_quantity": {
            "type": "number",
            "title": "Entry Quantity"
          },
          "exit_date": {
            "type": "string",
            "title": "Exit Date"
          },
          "exit_price": {
            "type": "number",
            "title": "Exit Price"
          },
          "exit_quantity": {
            "type": "number",
            "title": "Exit Quantity"
          },
          "realized_pnl": {
            "type": "number",
            "title": "Realized Pnl"
          },
          "realized_pnl_pct": {
            "type": "number",
            "title": "Realized Pnl Pct"
          },
          "holding_days": {
            "type": "integer",
            "title": "Holding Days"
          }
        },
        "type": "object",
        "required": [
          "trade_id",
          "ticker",
          "side",
          "entry_date",
          "entry_price",
          "entry_quantity",
          "exit_date",
          "exit_price",
          "exit_quantity",
          "realized_pnl",
          "realized_pnl_pct",
          "holding_days"
        ],
        "title": "ClosedTradeResponse",
        "description": "Closed trade response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderRequest": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "side": {
            "type": "string",
            "title": "Side",
            "description": "buy or sell"
          },
          "quantity": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Quantity"
          },
          "order_type": {
            "type": "string",
            "title": "Order Type",
            "description": "market, limit, stop, stop_limit",
            "default": "market"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "stop_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Price"
          },
          "limit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "side",
          "quantity"
        ],
        "title": "OrderRequest",
        "description": "Order request."
      },
      "OrderResponse": {
        "properties": {
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "side": {
            "type": "string",
            "title": "Side"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "order_type": {
            "type": "string",
            "title": "Order Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "filled_quantity": {
            "type": "number",
            "title": "Filled Quantity"
          },
          "filled_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filled Price"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "ticker",
          "side",
          "quantity",
          "order_type",
          "status",
          "filled_quantity",
          "filled_price",
          "created_at"
        ],
        "title": "OrderResponse",
        "description": "Order response."
      },
      "PortfolioResponse": {
        "properties": {
          "initial_cash": {
            "type": "number",
            "title": "Initial Cash"
          },
          "current_cash": {
            "type": "number",
            "title": "Current Cash"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/PositionResponse"
            },
            "type": "array",
            "title": "Positions"
          },
          "statistics": {
            "$ref": "#/components/schemas/PortfolioStatsResponse"
          },
          "closed_trades": {
            "items": {
              "$ref": "#/components/schemas/ClosedTradeResponse"
            },
            "type": "array",
            "title": "Closed Trades"
          }
        },
        "type": "object",
        "required": [
          "initial_cash",
          "current_cash",
          "positions",
          "statistics",
          "closed_trades"
        ],
        "title": "PortfolioResponse",
        "description": "Complete portfolio response."
      },
...
      "PositionResponse": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "entry_price": {
            "type": "number",
            "title": "Entry Price"
          },
          "current_price": {
            "type": "number",
            "title": "Current Price"
          },
          "cost_basis": {
            "type": "number",
            "title": "Cost Basis"
          },
          "market_value": {
            "type": "number",
            "title": "Market Value"
          },
          "unrealized_pnl": {
            "type": "number",
            "title": "Unrealized Pnl"
          },
          "unrealized_pnl_pct": {
            "type": "number",
            "title": "Unrealized Pnl Pct"
          },
          "is_long": {
            "type": "boolean",
            "title": "Is Long"
          },
          "is_short": {
            "type": "boolean",
            "title": "Is Short"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "quantity",
          "entry_price",
          "current_price",
          "cost_basis",
          "market_value",
          "unrealized_pnl",
          "unrealized_pnl_pct",
          "is_long",
          "is_short"
        ],
        "title": "PositionResponse",
        "description": "Position response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}